Update table content (316954)
As reported in WI#316954, the table featuring the parsing examples was not displaying the data correctly because the example and resulting data content contained pipes. GitHub does not permit | to be used within a table cell as part of the content. Pipes are only used to create table cells within GitHub - they cannot be used for data. Revised the content and substituted the html code for a vertical pipe (&#124;) so the content would display correctly.